Refactor stops part
