+++ busTrackerApi/src/main/kotlin/busTrackerApi/exceptions/BusTrackerException.kt
@@ -1,5 +1,7 @@
 package busTrackerApi.exceptions
 
+import io.ktor.websocket.*
+
 sealed class BusTrackerException : Exception() {
     class NotFound(override val message: String? = null) : BusTrackerException()
     class SoapError(override val message: String? = null) : BusTrackerException()
@@ -11,3 +13,5 @@ sealed class BusTrackerException : Exception() {
     class BadRequest(override val message: String? = null) : BusTrackerException()
     class Conflict(override val message: String? = null) : BusTrackerException()
 }
+
+class CloseSocketException(override val message: String? = null, val closeReason: CloseReason.Codes) : Exception()
