--- busTrackerApi/src/main/kotlin/busTrackerApi/extensions/SimpleJsonExtensions.kt
@@ -1,7 +1,0 @@
-package busTrackerApi.extensions
-
-import arrow.core.Either
-import busTrackerApi.exceptions.BusTrackerException.JsonError
-import simpleJson.exceptions.JsonException
-
-fun <T> Either<JsonException, T>.toBusTrackerException() = mapLeft { JsonError(it.message) }
